reduce code duplication and make genre clickable
