+++ lib/components/AlbumScreen/item_info.dart
@@ -6,6 +6,7 @@ import '../../models/jellyfin_models.dart';
 import '../../screens/artist_screen.dart';
 import '../../services/jellyfin_api_helper.dart';
 import '../../services/process_artist.dart';
+import '../icon_and_text.dart';
 import '../print_duration.dart';
 
 class ItemInfo extends StatelessWidget {
@@ -26,56 +27,17 @@ class ItemInfo extends StatelessWidget {
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
       children: [
         if (item.type != "Playlist") _ArtistIconAndText(album: item),
-        _IconAndText(
+        IconAndText(
             iconData: Icons.music_note,
             text: AppLocalizations.of(context)!.songCount(itemSongs)),
-        _IconAndText(
+        IconAndText(
             iconData: Icons.timer,
             text: printDuration(Duration(
                 microseconds:
                     item.runTimeTicks == null ? 0 : item.runTimeTicks! ~/ 10))),
         if (item.type != "Playlist")
-          _IconAndText(iconData: Icons.event, text: item.productionYearString)
+          IconAndText(iconData: Icons.event, text: item.productionYearString)
       ],
-    );
-  }
-}
-
-class _IconAndText extends StatelessWidget {
-  const _IconAndText({
-    Key? key,
-    required this.iconData,
-    required this.text,
-  }) : super(key: key);
-
-  final IconData iconData;
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 2),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(
-            iconData,
-            // Inactive icons have an opacity of 50% with dark theme and 38%
-            // with bright theme
-            // https://material.io/design/iconography/system-icons.html#color
-            color: Theme.of(context).iconTheme.color?.withOpacity(
-                Theme.of(context).brightness == Brightness.light ? 0.38 : 0.5),
-          ),
-          const Padding(padding: EdgeInsets.symmetric(horizontal: 2)),
-          Expanded(
-            child: Text(
-              text,
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-            ),
-          )
-        ],
-      ),
     );
   }
 }
@@ -94,7 +56,7 @@ class _ArtistIconAndText extends StatelessWidget {
           .getItemById(album.albumArtists!.first.id)
           .then((artist) => Navigator.of(context)
               .pushNamed(ArtistScreen.routeName, arguments: artist)),
-      child: _IconAndText(
+      child: IconAndText(
         iconData: Icons.person,
         text: processArtist(album.albumArtist, context),
       ),
